backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Machine learning
scikit-learn>=1.3.0

# MediaPipe for face and pose detection
mediapipe>=0.10.0

# OCR
pytesseract>=0.3.10

# PyTorch for MiDaS depth estimation
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# File watching
watchdog>=3.0.0
